Rename strategy lists file to LegacyPlaygroundResultStrategyLists
